2d wave  fsi pass
